core/alljoyn/alljoyn_java/src/org/alljoyn/bus/Version.java
/*
 * Copyright AllSeen Alliance. All rights reserved.
 *
 *    Contributed by Qualcomm Connected Experiences, Inc.,
 *    with authorization from the AllSeen Alliance, Inc.
 *    
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *    
 *        http://www.apache.org/licenses/LICENSE-2.0
 *    
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 *    
 *    Pursuant to Section 1 of the License, the work of authorship constituting
 *    a Work and any Contribution incorporated in the Work shall mean only that
 *    Contributor's code submissions authored by that Contributor.  Any rights
 *    granted under the License are conditioned upon acceptance of these
 *    clarifications.
 */

package org.alljoyn.bus;

public class Version {
    /**
     * Gives the version of AllJoyn Library
     *
     * @return the version of the AllJoyn Library
     */
    public static native String get();
    /**
     * Gives build information of AllJoyn Library
     *
     * @return build information of the AllJoyn Library
     */
    public static native String getBuildInfo();
    /**
     * Gives the version of AllJoyn Library as a single number
     *
     * @return the version of the AllJoyn Library as a single number
     */
    public static native int getNumeric();
}
